+++ server/lib/rdio-scanner/controller.js
@@ -21,16 +21,12 @@
 
 import { spawn, spawnSync } from 'child_process';
 import EventEmitter from 'events';
-import fs from 'fs';
-import path from 'path';
 import Sequelize from 'sequelize';
-import url from 'url';
 
 import { defaults } from './defaults.js';
 import { Log } from './log.js';
+import { version } from './version.js';
 import { WebSocket } from './websocket.js';
-
-const dirname = path.dirname(url.fileURLToPath(import.meta.url));
 
 const maxAuthenticationTries = 3;
 
@@ -42,6 +38,7 @@ const wsCommand = {
     livefeedMap: 'LFM',
     max: 'MAX',
     pin: 'PIN',
+    status: 'STS',
     ver: 'VER',
 };
 
@@ -79,7 +76,7 @@ export class Controller extends EventEmitter {
 
         this.tagsMap = this.getTagsMap();
 
-        this.version = getVersion();
+        this.version = version;
 
         this.websocket = null;
 
@@ -93,11 +90,11 @@ export class Controller extends EventEmitter {
     broadcastConfig() {
         if (this.websocket) {
             this.websocket.getSockets().forEach(async (socket) => {
-                if (this.isAccessRestricted && !socket.access) {
+                if ((this.isAccessRestricted && !socket.access) || socket.justStatus) {
                     return;
                 }
 
-                socket.send(JSON.stringify([wsCommand.pin]));
+                socket.send(JSON.stringify([wsCommand.config, this.getConfig(socket.scope)]));
             });
         }
     }
@@ -812,7 +809,30 @@ export class Controller extends EventEmitter {
                 socket.scope = this.getAccessScope();
             }
 
-            if (message[0] === wsCommand.ver) {
+            if (message[0] === wsCommand.status) {
+                if (typeof socket.status !== 'function') {
+                    socket.status = (call) => {
+                        if (socket.readyState !== 3) {
+                            socket.send(JSON.stringify([
+                                wsCommand.status,
+                                `${call.system},${call.talkgroup},${call.id}`,
+                            ]));
+
+                        } else {
+                            this.removeListener('call', socket.status);
+                            socket.status = undefined;
+                        }
+                    };
+                    this.addListener('call', socket.status);
+                    socket.send(JSON.stringify([wsCommand.status, true]));
+
+                } else {
+                    this.removeListener('call', socket.status);
+                    socket.status = undefined;
+                    socket.send(JSON.stringify([wsCommand.status, false]));
+                }
+
+            } else if (message[0] === wsCommand.ver) {
                 socket.send(JSON.stringify([wsCommand.ver, this.version]));
 
             } else if (this.isAccessRestricted && !socket.access && message[0] !== wsCommand.pin) {
@@ -857,7 +877,10 @@ export class Controller extends EventEmitter {
 
                         socket.livefeed = (call) => {
                             if (socket.readyState !== 3) {
-                                if (call.system in socket.scope && call.system in message[1]) {
+                                if (socket.justStatus) {
+                                    socket.send(JSON.stringify([wsCommand.status], `${call.system},${call.talkgroup},${call.id}`));
+
+                                } else if (call.system in socket.scope && call.system in message[1]) {
                                     if (socket.scope[call.system].includes(call.talkgroup)) {
                                         if (message[1][call.system] && message[1][call.system][call.talkgroup]) {
                                             socket.send(JSON.stringify([wsCommand.call, call]));
@@ -917,7 +940,8 @@ export class Controller extends EventEmitter {
                         socket.authCount = 0;
 
                         if (typeof socket.access.limit === 'number') {
-                            const count = Array.from(this.websocket.getSockets()).reduce((c, s) => s.access?.code === token ? ++c : c, 0);
+                            const count = Array.from(this.websocket.getSockets())
+                                .reduce((c, s) => (!s.justStatus && s.access?.code === token) ? ++c : c, 0);
 
                             if (count > socket.access.limit) {
                                 socket.send(JSON.stringify([wsCommand.max]));
@@ -1006,23 +1030,4 @@ export class Controller extends EventEmitter {
             })
         );
     }
-}
-
-function getVersion() {
-    const packageJsonFile = path.resolve(dirname, '../../package.json');
-
-    let version;
-
-    try {
-        const packageJson = JSON.parse(fs.readFileSync(packageJsonFile));
-
-        if ('version' in packageJson) {
-            version = packageJson.version;
-        }
-
-    } catch (error) {
-        console.error(error.message);
-    }
-
-    return version || 'unknown';
 }